+++ src/goupile/server/user.cc
@@ -1390,7 +1390,7 @@ void HandleChangeSecret(http_IO *io, const char *title)
         qr_ExportPng(qr, 0, &png);
 
         Span<const char> prefix = "data:image/png;base64,";
-        Size needed = prefix.len + (Size)sodium_base64_encoded_len(png.len, sodium_base64_VARIANT_ORIGINAL);
+        Size needed = prefix.len + (Size)sodium_base64_encoded_len((size_t)png.len, sodium_base64_VARIANT_ORIGINAL);
         Span<char> buf = AllocateSpan<char>(io->Allocator(), needed);
 
         CopyString(prefix, buf);
